test: update select-operator tests
