+++ bc_obps/registration/tests/endpoints/operator/test_request_admin_access.py
@@ -1,25 +1,14 @@
-from model_bakery import baker
 from registration.enums.enums import AccessRequestStates, AccessRequestTypes
-from registration.models import (
-    User,
-    UserOperator,
-)
+from registration.models import UserOperator
 from registration.tests.utils.bakers import operator_baker
 from registration.tests.utils.helpers import CommonTestSetup, TestUtils
 from registration.utils import custom_reverse_lazy
 
 
-class TestSelectOperatorRequestAccess(CommonTestSetup):
-    def test_request_subsequent_access_with_valid_payload(self, mocker):
+class TestSelectOperatorRequestAccessEndpoint(CommonTestSetup):
+    def test_request_admin_access_with_valid_payload(self, mocker):
         operator = operator_baker()
-        admin_user = baker.make(User, business_guid=self.user.business_guid)
-        baker.make(
-            UserOperator,
-            user=admin_user,
-            operator=operator,
-            role=UserOperator.Roles.ADMIN,
-            status=UserOperator.Statuses.APPROVED,
-        )
+        # We need to mock the send_operator_access_request_email function withing the application_access_service module
         mocked_send_access_request_email = mocker.patch(
             "service.application_access_service.send_operator_access_request_email"
         )
@@ -27,19 +16,20 @@ class TestSelectOperatorRequestAccess(CommonTestSetup):
             self,
             "industry_user",
             self.content_type,
-            {"operator_id": operator.id},
-            custom_reverse_lazy("request_access"),
+            {},
+            custom_reverse_lazy("request_admin_access", kwargs={"operator_id": operator.id}),
         )
-        response_json = response.json()
 
         # Assert that the email notification was sent
         mocked_send_access_request_email.assert_called_once_with(
             AccessRequestStates.CONFIRMATION,
-            AccessRequestTypes.OPERATOR_WITH_ADMIN,
+            AccessRequestTypes.ADMIN,
             operator.legal_name,
             self.user.get_full_name(),
             self.user.email,
         )
+
+        response_json = response.json()
 
         assert response.status_code == 201
         assert "user_operator_id" in response_json
@@ -49,6 +39,22 @@ class TestSelectOperatorRequestAccess(CommonTestSetup):
             user=self.user,
             operator=operator,
             status=UserOperator.Statuses.PENDING,
+            role=UserOperator.Roles.PENDING,
         ).exists()
 
         assert user_operator_exists, "UserOperator object was not created"
+
+    def test_request_access_with_invalid_payload(self):
+
+        response = TestUtils.mock_post_with_auth_role(
+            self,
+            "industry_user",
+            self.content_type,
+            {},
+            custom_reverse_lazy("request_admin_access", kwargs={"operator_id": 99999}),
+        )
+        assert response.status_code == 422
+        assert (
+            response.json().get("detail")[0].get("msg")
+            == "Input should be a valid UUID, invalid length: expected length 32 for simple format, found 5"
+        )
